Changelog: Fixed a problem with the calculator not working with decimal places

+++ python_calculator.py
@@ -25,7 +25,7 @@ class Calculator():
                 else:
                     self.new_num = False
                     display.insert(self.current_no,key)
-                self.current = int(display.get())
+                self.current = float(display.get())
                 print(self.current)
         except ValueError:
             if key == "+" or key == "-" or key == "*" or key == "/":
